Draft intermediate road waypoint computation between adjacent hamlets
This is a three step process:
  1. Determine which pairs of road endpoints are closest to each other.
  2. Compute the intermediate point between the two closest endpoints.
  3. Add a certain amount of noise to the intermediate point to make it
     more natural, this should be proportional to the distance between
     the two endpoints.

All this is done in this commit, I also
- Serialise + Deserialise road endpoint data properly
- Attempt to put in place guards to protect against crashes

Crashes still happen! Unfortunately. Also, LSystemVillageLayer's
constructor is sometimes not even called, I think due to hacks with the
SimpleProfiler's GetStatus method, in particular introducing a boolean
check `isLoggingStatus` and running

```
if (isLoggingStatus)  {
	return "SomeStatus";
}
```

This doesn't work, I think something I am doing by deferring the call to
the LSystemVillageLayer LayerDependency is breaking things. Will seek to
investigate + attempt to address in next commit.