Wrangle top level vars + implement approach to propagate to deps
- Rename ApplyArgument to LayerPostmount
- Signal pattern for PlayLayer
- Some pre-work regarding global args
- Ensure that we use the arguments that are defined in our scene, no
  hardcoding of values in the script ... except for the keys of course
  ...

There are issues with this, the whole motivation of passing these
variables is to determine how much to compute and how much to render,
and it seems that it is running _after_ we do those things.

One workaround for this is to have conservative initial computation +
render values, then on the first timer timeout tick we can use the
values concerning where we are.

Still not quite what we need though, we do need to provide the context
of where the camera is before we start constructing anything.

Tricky! Needs a bit more careful thought but will retain this for now.

---
PostScript

Seems like I need a way to modify what I pass to the constructor of a
Layer from the GenerationSource ... hmm