WIP, bump to terrain3d 0.9.3a
Terrain3D 0.9.1 doesn't support Godot 4.4.1 [3], there are issues with the
camera view fulstrum not being resolved properly [1][2]. Terrain 0.9.3a
_does_ support Godot 4.4.1 however [4].

This results in blue artefacts as documented in this issue [7], which
detracts from the user experience.

So upgrading seems like a good idea, which may fix.

To upgrade, we need to follow this recipe [5], in particular:

- Build Terrain3D tag 0.9.3 from source, with the 4.4.1 branch of
  godot-cpp checked out in the relevant git submodule [6]
- Clone this generator:
  https://github.com/Delsin-Yu/CSharp-Wrapper-Generator-for-GDExtension
- Delete all folders in addons for this project except for gdunit4 +
  cs_wrapper_generator_for_gde
- Copy the terrain_3d folder in addons (including the built dlls for your
  desired platform(s)) to the CSharp-Wrapper-Generator-for-GDExtension project
- Enable terrain3d, gdunit4 + cs_wrapper_generator_for_gde
- Click the generate tab (next to the file manager in engine) + click
  generate
- Rename to terrain_3d_csharp_delsin_yu_generator + rename namespace
  GDExtension.Wrappers to namespace TokisanGames
- Copy said generated folder to this project, along with the new
  terrain3d folder for tag 0.9.3a, deleting the old terrain3d folder for
  0.9.1

Note this doesn't quite work, the old terrain_3d_csharp folder seems to
be being used by this project in various ways, so there is an additional
gap to close ... possibly by altering the way terrain_3d is invoked.

Maybe this can be done in a follow-up commit.

For now I will retain the old terrain_3d_csharp folder, alongside the
new terrain_3d_csharp_delsin_yu_generator folder.

[1] https://github.com/Spinning-Cat-Studios/LayerProcGenExampleProject/issues/3#issuecomment-2906778650
[2] https://chatgpt.com/share/6831a9e3-e1e0-8005-9fe9-61375d859b85
[3] https://github.com/TokisanGames/Terrain3D/releases/tag/v0.9.1-beta
[4] https://github.com/TokisanGames/Terrain3D/releases/tag/v0.9.3-beta
[5] https://github.com/TokisanGames/Terrain3D/blob/42569d0c42e06a7c0b4df7136347c3783890c585/doc/docs/generating_csharp_bindings.md
[6] https://terrain3d.readthedocs.io/en/stable/docs/building_from_source.html
[7] https://github.com/Spinning-Cat-Studios/LayerProcGenExampleProject/issues/3