Rework wrapper to allow game to run without crashing
Terrain3D object is not rendering though, not sure why that is, also the
road painter service seems to think the object is null.

The latter reason is because we're not using a NodePath reference to the
Terrain3D within the Layer Argument Dictionary, this should admit an
easy fix.

Not sure about the former, this bears further thought.