WIP, changes to support subtypes in CBDL, play layer + landscape layer
Note, additional changes required to AbstractDataLayer as s_LayerDict
does not recognise subtypes for a given type; the current structure only
supports one subtype per type, which is not the intent.

To fix additional changes are required at the lower level of the code,
or alternatively we need to consider splitting LocationLayer into
LocationLayerA ... LocationLayerD, CultivationLayer to CultivationLayerA
through CultivationLayerD, etc, which I don't think is ideal.