Add player path + lazy evaluation timeout values to dictionary
The idea is that the player path information will be used to derive, on
construction, whether a node in the scenegraph should be constructed
and/or rendered

The lazy evaluation timeout indicates how frequently to re-evaluate
distances relative to the player position, since evidently the player
might move around in the scene over time

An alternative + perhaps more elegant approach might be to track the
player position within the player controller, and upon threshold being
reached in terms of player delta, to trigger a re-evaluation

That might be better, will think about this.

But for now I'll retain the lazy evaluation timeout dictionary value in
case I decide to use it after all.