Implement signal to emit if camera position exceeds threshold
Send a signal via SignalBus to indicate that nodes should be
reconstructed

There is some finesse required here, we don't want to reconstruct all
nodes every time this condition is met, just those that are on the
boundary, i.e. remove nodes that are too far away, and add nodes that
have come into range.

This can probably be done by querying sqlite for nodes that are within a
certain distance and circumventing the constructors for those nodes, in
lieu of focusing only on nodes that are not in the database.

Regardless, now of this logic should be implemented here, this is just a
signal, the consumer of this signal should handle the logic of
reconstructing nodes as indicated above.