Fix race condition /w freelookcamera, return early if villagelayer done
In order to lazily load properly, we will need to implement lazy
evaluation for the village layer properly. This will be done in a
follow-up commit.

Another thing worth checking is that ReconstructNodes is performing its
function, in that

- Chunks out of range coming into range are now rendered
- Chunks that _were_ in range going out of range are removed
- Chunks that remain in range are not re-computed

At present I'm not sure that this is truly occurring, I do know that
ReconstructNodes is being emitted + can be consumed, so that is working,
and I know that I'm not spamming emission of the signal, there is just
the small matter of ensuring that

- Measurements are correct, i.e. we're putting the right things in the
  right boxes to evaluate relative xz distance,
- That checkpoints are being updated correctly, i.e. that we're not just
  measuring relative to the same thing again + again,
- That distance to a chunk is being measured relative to the camera xz
  position properly (maybe repetition of the above but worth clarifying)
- That positions of chunks are measured correctly, i.e. that I'm using
  the right coordinates for them.

So a bit more to do, but progress.