Comment out playlayer rebuild, fix early return borking load
- In EnsureLoadedInBounds I had an early return which caused the game to
  terminate / sigterm, so I've removed it
- In EnsureLoadedInBounds I had unreachable logic, so I've improved the
  control flow
- Within the LazyEvaluationHandler I was rebuilding the PlayLayer, this
  was a very expensive operation which I've commented out for now
- Within LandscapeLayerOrchestrator I was passing `null` to
  EnsureLoadedInBounds, now fixed

There is a follow-up issue where the referencePosition within
UpdateLayerBasedOnCameraMovement in the LandscapeLayerOrchestrator seems
always to be (0,0,0), this will require fixing, which will be done in a
follow-up commit.

Likely this is due to potentially passing a node that doesn't change
when the camera moves, or simply that the player camera node isn't being
found properly, and the value is falling back to a default.